Return MONITOR_INVALID_VALUE if too many mailboxes are requested
Mailboxes and the enclave metadata must fit in one 256MB region.

Previously, security monitor was checking that the number of pages
fits in one region, but if too large a value is provided for
num_mailboxes, the number of pages calculation overflows, enabling an
attacker to provide a value that passes the check on number of pages
while having too many mailboxes configured. This allows the attacker
to overlay mailboxes on other structures.

This adds a check that the number of mailboxes will not cause the
number of pages to exceed 256M/PAGE_SIZE.