Fix memcpy() invocation in mail_send and mail_receive
phys_ptr_t is represented as an unsigned long, I guess because in
kernel space physical pointers are not typically dereferenceable.

However, this leads to confusion, because the C compiler says it is a
type error to use a phys_ptr_t as a source our destination in
memcpy(), which lead to using the address of the address of the
mailbox rather than the address of the mailbox.

Added a cast and removed the extra "&".